change client API error to payment declined error - select specific error codes

+++ src/Core/PaymentGateway.Shared/Models/Controller/Requests/PostPaymentRequest.cs
@@ -19,7 +19,7 @@ public class PostPaymentRequest
     public required string Currency { get; set; }
     [Required]
     [JsonPropertyName("amount")]
-    public required int Amount { get; set; }
+    public required long Amount { get; set; }
     [Required]
     [JsonPropertyName("cvv")]
     public required string Cvv { get; set; }
